% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edmonds_karp.R
\name{edmonds_karp}
\alias{edmonds_karp}
\title{edmonds_karp: Edmonds-Karp Algorithm for Maximum Flow}
\usage{
edmonds_karp(arr, source, sink)
}
\arguments{
\item{arr}{The adjacency matrix representing the network.}

\item{source}{The index of the source node.}

\item{sink}{The index of the sink node.}
}
\value{
A list containing the final flow matrix and the value of the maximum flow.
}
\description{
This function implements the Edmonds-Karp algorithm to find the maximum flow in a network represented by an adjacency matrix.
It utilizes the Breadth-First Search (BFS) algorithm and the get_flow_min function to iteratively find augmenting paths and update flow values.
}
