% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{edmondsKarp}
\alias{edmondsKarp}
\title{Edmonds-Karp algorithm for maximum flow}
\usage{
edmondsKarp(capacity, source, sink)
}
\arguments{
\item{capacity}{A NumericMatrix where the entry at (i, j) is the capacity of the edge from node i to node j.}

\item{source}{The index of the source node in the flow network (using 1-based indexing as in R).}

\item{sink}{The index of the sink node in the flow network (using 1-based indexing as in R).}
}
\value{
A list containing two elements: the flow matrix and the value of the maximum flow.
}
\description{
This function implements the Edmonds-Karp algorithm to find the maximum flow in a flow network.
}
