% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ford_fulkerson150}
\alias{ford_fulkerson150}
\title{Main function to implement Ford-Fulkerson algorithm}
\usage{
ford_fulkerson150(graph, source, sink)
}
\arguments{
\item{source}{The index of the source node in the flow network (using 1-based indexing as in R).}

\item{sink}{The index of the sink node in the flow network (using 1-based indexing as in R).}

\item{graph}{graph is the adjacency matrix of a graph where the entry at (i, j) is the capacity of the edge from node i to node j.}
}
\value{
It returns a list containing the flow matrix called residual matrix ( the way the flow should propagate) and the value of the maximum flow from the source.
}
\description{
Ford Fulkerson Algorithm to find maximum flow of a graph using a backtracking method to find the augmented path, The function to find augmented path is defined inside the main function and its callled find_augmenting_path. Note that Augmented path is recursively founded
}
